configure nestJs and vue SSR
