+++ package.json
@@ -8,7 +8,6 @@
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint --fix",
     "test": "vue-cli-service test",
-    "server:ssr:dev": "rimraf dist && cross-env NODE_ENV=production ts-node --project tsconfig.webpack.json server/SSR/main.ts",
     "server:dev": "tsnd --respawn --debug server/main.ts",
     "server:start": "cross-env NODE_ENV=production ts-node server/main.ts",
     "ssr:dev": "ts-node --project tsconfig.webpack.json ssrDev/ssrDevServer.ts",
